MB-68266: Use size_t for task/taskable counts in executor
Change numTaskables(), numTasksForOwner() return types and
numWorkers storage from int to size_t to match the return type
of container.size(). This eliminates implicit narrowing
conversions (64-bit size_t to 32-bit int) caught by
-Werror=shorten-64-to-32.

Using size_t ensures type correctness when working with
container sizes and avoids potential overflow issues with
large counts.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/244288
Tested-by: Build Bot <build@couchbase.com>

+++ executor/folly_executorpool.cc
@@ -593,7 +593,7 @@ struct FollyExecutorPool::State {
     }
 
     /// @returns the number of taskables registered.
-    int numTaskables() const {
+    size_t numTaskables() const {
         return taskOwners.size();
     }
 
@@ -601,7 +601,7 @@ struct FollyExecutorPool::State {
      * Returns the number of tasks owned by the specified taskable, plus
      * any pending reset after being cancelled.
      */
-    int numTasksForOwner(const Taskable& taskable) {
+    size_t numTasksForOwner(const Taskable& taskable) {
         auto& owner = taskOwners.at(&taskable);
         return owner.locator.size() + owner.pendingTaskResets;
     }
@@ -944,7 +944,7 @@ void FollyExecutorPool::unregisterTaskable(Taskable& taskable, bool force) {
 
 size_t FollyExecutorPool::getNumTaskables() const {
     cb::NoArenaGuard guard;
-    int numTaskables = 0;
+    size_t numTaskables = 0;
     futurePool->getEventBase()->runImmediatelyOrRunInEventBaseThreadAndWait(
             [state = this->state.get(), &numTaskables] {
                 numTaskables = state->numTaskables();
